Refactored Radio in Typescript
