+++ src/components/Radio/index.tsx
@@ -1,0 +1,3 @@
+import Radio from './Radio';
+
+export default Radio;
